More to a chroot based build
This allows the build process to be much more reliable, with the tradeoff of mandating root to be able to function properly. This fixes stability issues that I had between Github Codespaces and my local machine, and make self tests finally run properly.

A CI shall be set up soon to make sure the build process still works and selftests still pass.